warning

+++ src/threadPool/threadPool.h
@@ -15,7 +15,7 @@ public:
   }
   ~ThreadGuard()
   {
-    for (int i = 0; i < threads.size(); i++)
+    for (size_t i = 0; i < threads.size(); i++)
     {
       if (threads[i].joinable())
       {
